fix minor bug

+++ TMessagesProj/src/main/java/org/telegram/ui/Cells/ChatMessageCell.java
@@ -653,7 +653,7 @@ public class ChatMessageCell extends BaseCell implements SeekBar.SeekBarDelegate
     private boolean drawEditButton;
     private boolean editPressed;
     private int editStartX;
-    private int editStartY;
+    private float editStartY;
     //Devgram/
     private int layoutWidth;
     private int layoutHeight;
